+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#193](https://github.com/os2display/display-api-service/pull/193)
+  - Adds support for interactive slides.
+  - Adds interactivity for creating quick bookings from a slide through Microsoft Graph.
+  - Adds KeyVaultService that can serve key-value entries from the environment for storing secrets.
+
 ## [2.0.6] - 2024-06-28
 
 - [#208](https://github.com/os2display/display-api-service/pull/208)